+++ include/ui_settings_display_sound.h
@@ -83,6 +83,7 @@ class DisplaySoundSettingsOverlay : public OverlayBase {
     void handle_time_format_changed(int index);
     void handle_animations_changed(bool enabled);
     void handle_system_keyboard_changed(bool enabled);
+    void handle_keep_navbar_changed(bool enabled);
 
     // Display handlers
     void handle_dark_mode_changed(bool enabled);
@@ -190,6 +191,7 @@ class DisplaySoundSettingsOverlay : public OverlayBase {
     static void on_time_format_changed(lv_event_t* e);
     static void on_animations_changed(lv_event_t* e);
     static void on_system_keyboard_changed(lv_event_t* e);
+    static void on_keep_navbar_changed(lv_event_t* e);
 
     // Display
     static void on_dark_mode_changed(lv_event_t* e);